Refactor create PErson
